add more metadata to the Tree and the storage path
Not only do stored modules need to know their version if it exists, but
any relative source needs to know all the ancestor versions in order to
resolve correctly.

This is still adding to the overloaded Load method, since that code
paths is already understood. Breaking that up into more manageable
chunks can be done later.